+++ trading-engine/engine/trade.go
@@ -6,10 +6,10 @@ import (
 )
 
 type Trade struct {
-	TakerOrderID uint64 `json:"taker_order_id"`
-	MakerOrderID uint64 `json:"maker_order_id"`
-	Amount       uint64 `json:"amount"`
-	Price        uint64 `json:"price"`
+	TakerOrderID uint64  `json:"taker_order_id"`
+	MakerOrderID uint64  `json:"maker_order_id"`
+	Amount       uint64  `json:"amount"`
+	Price        float64 `json:"price"`
 }
 
 func (trade *Trade) FromJSON(msg []byte) error {
@@ -17,7 +17,7 @@ func (trade *Trade) FromJSON(msg []byte) error {
 }
 
 func (trade Trade) String() string {
-	return fmt.Sprintf("{TakerOrderID:%d, MakerOrderID:%d, Amount:%d, Price:%d} ", trade.TakerOrderID, trade.MakerOrderID, trade.Amount, trade.Price)
+	return fmt.Sprintf("{TakerOrderID:%d, MakerOrderID:%d, Amount:%d, Price:%f} ", trade.TakerOrderID, trade.MakerOrderID, trade.Amount, trade.Price)
 }
 
 func (trade *Trade) ToJSON() []byte {
